Add checks for function existence and empty params
Added a check for the existence of is_wp_rocket_crawling() before calling it to prevent errors. Updated inline_css() to return early if the URL is empty, and set a default value for the $url parameter. Also, set $value to an empty string if get_field returns null in SaltBase::get_acf_option().